build with libdevmapper
The reason was simple: the installer would fail against md raid1 root:

  Can't install grub on /dev/sda :Path `/boot/grub'
  is not readable by GRUB on boot. Installation is
  impossible. Aborting.

When diagnosing,

  chroot# /usr/sbin/grub-probe -t fs /boot/grub
  /usr/sbin/grub-probe: error: disk `hostdisk//dev/md/md1' not found.

...and with -v it's clear that /dev/md -> /dev/evms/md symlink
(a EVMS related hack in alterator-grub as of today) is unpassable
for grub-probe which seems to be the intended behaviour for avoiding
hitting device duplicates; see also [1].  Of course /dev/md/md1 is
alive and kickin'.

Long story short, enabling devmapper support (which might come handy
for LUKS either) and replacing a symlink with a bindmount solved this.

[1] http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=550704