spec: handle legacy /usr/lib/grub symlink
As noted in #27935, ALT's grub-0.9x had /usr/lib/grub -> /boot/grub
symlink (the package was i586 only due to the lack of -m32 here and
then).  RPM4 doesn't handle directory<->symlink transformations
exactly flawlessly, that is, /boot/grub of the *new* package
would get wiped in the end requiring an apt-get reinstall.

Thanks serpiph@ for spotting and reporting this.