implemented explicit bootloader id specification
The problem solved is that the bootloader_id used to form
/boot/efi/EFI/ subdir name is hardcoded to be deduced from
GRUB_DISTRIBUTOR with a pretty arbitrary limitation, namely
"get only the very first word downcased", as the only rule.

"alt" is a bit too terse and kludging this "solution"
was looking too bad already so I just went forward
to introduce one more variable: GRUB_BOOTLOADER_ID.

Thus there's a pretty label for EFI Boot Manager along with
a reasonable ESP path component; a fallback in grub-install
is also there.

The sysconfig file path was incorrect for ALT, even if that's
mitigated by a recently added symlink.

Unfortunately all of this became pretty intertangled :-/