sort kernels by mtime instead of ctime
Sorting kernels by ctime is less reliable and may cause some time issues,
when a recent kernel has ctime less than older kernel on a machine. This
is possible because ctime of a kernel depends on the machine's time during
the kernel update. mtime of a kernel is independent of the machine's time
and store building time, which is exactly what we need.